Corrected Error in FDD_r, Updated Help
FDD_r had a small error in testing for empty lists that made it have a
chance of hitting an error on disconnected graphs (e.g. General
Relativity broke it).

Fixed this by putting in several test conditions on the lists we work
on, and defaulting to a random, new, node if we can't pick a node
deterministically.

I also made a small update to the "print algorithms" text, since I had
to go back and look at the code to find correct parameters for
restarting!