Revisions for Multilayered Networks/outputs
Added an extra command line argument for having a "verbose" algorithm
output. That is, you can now include a "myprint" in an algorithm file,
that will print/save whatever you want. Specifically, this is designed
to support working with networks or inference concepts that are not
directly supported in the main graphinfer.py file.